positions_dashboard/requirements.txt
streamlit
pandas
psycopg2-binary
requests
streamlit-autorefresh